// src/components/specific/smartchart/enum.ts
export enum Formula {
	COUNT = "COUNT",
	AVG = "AVG",
	SUM = "SUM",
	MIN = "MIN",
	MAX = "MAX"
}
export enum ChartTypes {
	line = "line",
	bar = "bar",
	pie = "pie",
	num = "num",
	radar = "radar",
	panel = "panel"
}
export enum ViewTypes {
	system = "system",
	custom = "custom",
	company = "company"
}
export enum ColumnTypes {
	number = "number",
	string = "string"
}
export enum SaveTypes {
	save = "save",
	saveAs = "saveAs"
}
export enum ActionTypes {
	save = "save",
	init = "init",
	changeEditView = "changeEditView"
}

export enum ChartAxis {
	yAxis = "yAxis",
	color = "color"
}